Fix policy manager loading
The policy manager was also failing to load because
it was using webinos-utilities, which was not present
in node_modules for this module.  This fixes that
problem.  It also automatically creates a new, valid
policy file when the policy manager starts, to get rid
of another error.

Jira issue: WP-968