Add asynchronous prompting support.
A unique token is generated for each prompt, and the prompt request details and callback are cached with this token. The dashboard prompting app responds to prompts using this token, which is then used to call the appropriate callback.

Also rationalised the prompt choices to be the same for all prompt types.

Jira issue: WP-956